+++ src/lib.rs
@@ -4,4 +4,4 @@ pub mod c_nodes;
 pub mod command;
 pub mod model;
 mod mount;
-mod walk;
+mod make;